requirements.txt
django==4.1.1
djangorestframework
psycopg2>=2.8
requests
beautifulsoup4
django-enumchoicefield
django-cron
django-cors-headers
Pillow
django-allauth
dj-rest-auth
django-cleanup